third_party/WebKit/LayoutTests/fast/dom/shadow/querySelector-for-useragent-shadowroot.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
  <form id="form"><input id="input" type="text" name="text"></form>
  <details id="details">
    <summary id="summary"></summary>
    <p id="p"></p>
  </details>
  <meter value="2" min="0" max="10" id="meter"></meter>
  <progress value="22" max="100" id="progress"></progress>
  <pre id="console"></pre>
</body>
<script>
description('crbug.com/337616: test for querySelectorAll with ::shadow and /deep/');

shouldBe('document.querySelectorAll("form /deep/ *").length', '1');
shouldBe('document.querySelectorAll("form /deep/ *")[0].id', '"input"');
shouldBe('document.querySelectorAll("form::shadow *").length', '0');
shouldBe('document.querySelectorAll("input /deep/ *").length', '0');
shouldBe('document.querySelectorAll("input::shadow *").length', '0');
shouldBe('document.querySelectorAll("details /deep/ *").length', '2');
shouldBe('document.querySelectorAll("details /deep/ *")[0].id', '"summary"');
shouldBe('document.querySelectorAll("details /deep/ *")[1].id', '"p"');
shouldBe('document.querySelectorAll("details::shadow *").length', '0');
shouldBe('document.querySelectorAll("summary /deep/ *").length', '0');
shouldBe('document.querySelectorAll("summary::shadow *").length', '0');
shouldBe('document.querySelectorAll("meter /deep/ *").length', '0');
shouldBe('document.querySelectorAll("meter::shadow *").length', '0');
shouldBe('document.querySelectorAll("progress /deep/ *").length', '0');
shouldBe('document.querySelectorAll("progress::shadow *").length', '0');
</script>
</html>
